Added input_validators.py. Moved all input validators & password strength verifiers to new .py file. Added logout route with basic return string.
